Refactor: extract sim host shared base + c_api glue (-2800 lines)
Mirror the onboard host refactor (#880…#928) on the sim path. Pulls
~2800 duplicate lines out of src/{a2a3,a5}/platform/sim/host/ into a
shared SimDeviceRunnerBase + c_api_shared.cpp + memory_allocator.cpp
under src/common/platform/sim/host/. Per-arch DeviceRunner keeps only
the bits that genuinely differ between a2a3 and a5 sim:

- aicore_execute signature (a5 has extra aicore_pmu_ring_addrs arg)
- dlsym'd function-pointer table (a2a3 has dep_gen / pmu_reg_addrs /
  aicore_rotation_table setters; a5 doesn't)
- init_* alloc strategy (a2a3 uses mem_alloc_ via captured lambdas;
  a5 uses std::malloc via prof_alloc_cb static functions — preserved
  as-is, no behavior change)
- finalize() collector semantics (a2a3 releases shm to mem_alloc_
  per-run; a5 stop()s per-run, full finalize at run-end via
  prof_free_cb — preserved as-is)
- run() middle (dep_gen gating on a2a3 only; different SIM_REG_*
  constants)

SimDeviceRunnerBase hosts the byte-identical methods + their state:
setup_static_arena, acquire_pooled_*, create_thread,
attach_current_thread, allocate_tensor / free_tensor / copy_*,
register_callable[_host_orch], unregister_callable, has_callable,
bind_callable_to_runtime, prepare_orch_so, upload_chip_callable_buffer,
print_handshake_results, release_callable_state, ensure_device_initialized,
set_*_enabled / output_prefix accessors, last_device_wall_ns, the
shared mem_alloc_ / gm_*_arena_ / callable maps / chip_callable
buffer pool / collector instances / kernel_args_ / device_wall ptr /
log/dlopen counters.

Mechanical-fix: setup_static_arena standardized to "release all
arenas on any failure" (matches a2a3 sim + onboard PR #922). a5 sim
had been keeping earlier-committed peers alive on later-region
failure; the new common impl drops that to match the onboard
invariant.

Latent-fix carried over from onboard #928: c_api_shared's
run_prepared wraps the placement-new'd Runtime in RAIIScopeGuard so
its dtor fires on every exit path (manual r->~Runtime() in the prior
sim c_api was bypassed by catch(...) on exception).

Polymorphism via SimDeviceRunnerBase virtuals: ~SimDeviceRunnerBase
(public), run(), finalize(), set_dep_gen_enabled() (default no-op,
a2a3 overrides). c_api_shared.cpp works through
SimDeviceRunnerBase * and dispatches via those virtuals — same
pattern as the onboard #928 split.

Per-arch pto_runtime_c_api.cpp shrinks from ~420 lines to ~55 (just
create_device_context + ACL stubs). memory_allocator.cpp was
byte-identical, deleted from both arch subdirs and lives once in
common/.

Both arches built clean. nm -D verifies all 17 ChipWorker dlsym
targets exported on both sim libhost_runtime.so. ST passes 38/38 on
a2a3sim L1+L2 (devices 0,1) and 22/22 on a5sim L1+L2 (devices 0,1);
examples spot-checked (scalar_data_test, vector_example,
benchmark_bgemm on a2a3sim; vector_example, bgemm on a5sim).
