+++ src/a2a3/platform/sim/host/device_runner.cpp
@@ -9,39 +9,34 @@
  * -----------------------------------------------------------------------------------------------------------
  */
 /**
- * Device Runner Implementation - Thread-Based Simulation
- *
- * This file implements the simulated device execution using host threads.
- * It provides the same API as the real a2a3 implementation but uses
- * std::thread instead of CANN runtime APIs.
- *
- * aicpu_execute and aicore_execute_wrapper are loaded dynamically via dlopen from
- * the binaries passed to launch_runtime.
- *
- * Cross-platform notes:
- * - Linux: Uses MAP_ANONYMOUS for anonymous memory mapping
- * - macOS: Uses MAP_ANON (aliased) and MAP_JIT for executable memory on Apple Silicon
- *   which requires W^X (write xor execute) protection toggling via pthread_jit_write_protect_np
+ * a2a3 sim DeviceRunner implementation — wired against a2a3's aicore_execute
+ * signature + dep_gen support. Shared arena/tensor/callable lifecycle lives
+ * on SimDeviceRunnerBase; see device_runner_base.cpp.
  */
 
 #include "device_runner.h"
 
+#include <dlfcn.h>
 #include <stdlib.h>
 #include <sys/stat.h>
+#include <unistd.h>
 
 #include <atomic>
 #include <chrono>
 #include <cstdio>
+#include <cstring>
+#include <functional>
 #include <string>
 #include <vector>
 
 #include "aicpu/platform_aicpu_affinity.h"
-#include "callable.h"
 #include "callable_protocol.h"
-#include "chip_callable_layout.h"
+#include "common/memory_barrier.h"
+#include "common/platform_config.h"
+#include "common/unified_log.h"
 #include "cpu_sim_context.h"
 #include "host/raii_scope_guard.h"
-#include "utils/elf_build_id.h"
+#include "runtime.h"
 
 // dep_gen_replay_emit_deps_json: strong symbol provided by
 // runtime/tensormap_and_ringbuffer/host/dep_gen_replay.cpp when that runtime is
@@ -59,177 +54,14 @@ extern "C" __attribute__((weak, visibility("hidden"))) int dep_gen_replay_emit_d
     return -1;
 }
 
-// Function pointer types for dynamically loaded executors
-typedef int (*aicpu_execute_func_t)(Runtime *runtime);
-typedef void (*aicore_execute_func_t)(
-    Runtime *runtime, int block_idx, CoreType core_type, uint32_t physical_core_id, uint64_t regs,
-    uint32_t enable_profiling_flag, uint64_t l2_swimlane_aicore_rotation_table
-);
-typedef void (*set_platform_regs_func_t)(uint64_t regs);
-typedef void (*set_platform_dump_base_func_t)(uint64_t dump_data_base);
-typedef void (*set_dump_tensor_enabled_func_t)(bool enable);
-typedef void (*set_platform_pmu_base_func_t)(uint64_t pmu_data_base);
-typedef void (*set_pmu_enabled_func_t)(bool enable);
-
-namespace {
-
-bool write_all_bytes(int fd, const uint8_t *data, size_t size) {
-    size_t total_written = 0;
-    while (total_written < size) {
-        ssize_t written = write(fd, data + total_written, size - total_written);
-        if (written <= 0) {
-            return false;
-        }
-        total_written += static_cast<size_t>(written);
-    }
-    return true;
-}
-
-bool create_temp_so_file(const std::string &path_template, const uint8_t *data, size_t size, std::string *out_path) {
-    std::vector<char> path_buf(path_template.begin(), path_template.end());
-    path_buf.push_back('\0');
-
-    int fd = mkstemp(path_buf.data());
-    if (fd < 0) {
-        return false;
-    }
-
-    // dlopen requires the file to be executable; mkstemp creates 0600 (no exec bit)
-    if (fchmod(fd, 0755) != 0) {
-        close(fd);
-        unlink(path_buf.data());
-        return false;
-    }
-
-    bool ok = write_all_bytes(fd, data, size);
-    if (close(fd) != 0) {
-        ok = false;
-    }
-    if (!ok) {
-        unlink(path_buf.data());
-        return false;
-    }
-
-    *out_path = path_buf.data();
-    return true;
-}
-
-}  // namespace
-
-// =============================================================================
-// DeviceRunner Implementation
-// =============================================================================
-
 DeviceRunner::~DeviceRunner() { finalize(); }
-
-int DeviceRunner::setup_static_arena(size_t gm_heap_size, size_t gm_sm_size, size_t runtime_arena_size) {
-    // Three independent device_malloc'd buffers: GM heap, PTO2 SM, prebuilt
-    // runtime arena. Split out from a single large allocation because the
-    // combined size can exceed the device allocator's largest contiguous
-    // block. Each arena commits exactly one region, so its base() is the
-    // pooled pointer the caller wants.
-    //
-    // Idempotent for the production case (sizes do not change across a
-    // worker's lifetime). If a caller asks for a larger layout on any
-    // region, redo just that region.
-    auto commit_region = [](DeviceArena &arena, size_t &cached_size, size_t requested_size) -> int {
-        if (requested_size == 0) {
-            if (arena.is_committed() && cached_size != 0) {
-                arena.release();
-                cached_size = 0;
-            }
-            return 0;
-        }
-        if (arena.is_committed() && requested_size <= cached_size) {
-            return 0;
-        }
-        arena.release();
-        cached_size = 0;
-        arena.reserve(requested_size, DeviceArena::kDefaultBaseAlign);
-        if (arena.commit(DeviceArena::kDefaultBaseAlign) == nullptr) {
-            arena.release();
-            return -1;
-        }
-        cached_size = requested_size;
-        return 0;
-    };
-    if (commit_region(gm_heap_arena_, cached_gm_heap_size_, gm_heap_size) != 0) return -1;
-    if (commit_region(gm_sm_arena_, cached_gm_sm_size_, gm_sm_size) != 0) {
-        gm_heap_arena_.release();
-        cached_gm_heap_size_ = 0;
-        return -1;
-    }
-    if (commit_region(runtime_arena_pool_, cached_runtime_arena_size_, runtime_arena_size) != 0) {
-        gm_heap_arena_.release();
-        gm_sm_arena_.release();
-        cached_gm_heap_size_ = 0;
-        cached_gm_sm_size_ = 0;
-        return -1;
-    }
-    return 0;
-}
-
-void *DeviceRunner::acquire_pooled_gm_heap() {
-    if (!gm_heap_arena_.is_committed()) return nullptr;
-    return gm_heap_arena_.base();
-}
-
-void *DeviceRunner::acquire_pooled_gm_sm() {
-    if (!gm_sm_arena_.is_committed()) return nullptr;
-    return gm_sm_arena_.base();
-}
-
-void *DeviceRunner::acquire_pooled_runtime_arena() {
-    // hbg calls setup_static_arena(...,0) and leaves runtime_arena_pool_
-    // uncommitted — fail loudly if a caller asks for it anyway.
-    if (!runtime_arena_pool_.is_committed()) return nullptr;
-    return runtime_arena_pool_.base();
-}
-
-std::thread DeviceRunner::create_thread(std::function<void()> fn) {
-    int dev_id = device_id_;
-    return std::thread([dev_id, fn = std::move(fn)]() {
-        pto_cpu_sim_bind_device(dev_id);
-        fn();
-        pto_cpu_sim_bind_device(-1);
-    });
-}
-
-int DeviceRunner::attach_current_thread(int device_id) {
-    if (device_id < 0) {
-        LOG_ERROR("Invalid device_id: %d", device_id);
-        return -1;
-    }
-    if (device_id_ != -1 && device_id_ != device_id) {
-        LOG_ERROR(
-            "DeviceRunner already initialized on device %d; finalize before switching to device %d", device_id_,
-            device_id
-        );
-        return -1;
-    }
-
-    // Per-thread bind so sim hooks (TPUSH/TPOP, identity helpers) route through
-    // the correct context. acquire is process-wide and idempotent (no-op after
-    // first call for a given device_id), so it is safe to fold in here.
-    pto_cpu_sim_bind_device(device_id);
-    pto_cpu_sim_acquire_device(device_id);
-    device_id_ = device_id;
-    return 0;
-}
-
-int DeviceRunner::ensure_device_initialized() {
-    // device_id_ was set in attach_current_thread() during simpler_init.
-    int rc = attach_current_thread(device_id_);
-    if (rc != 0) return rc;
-    return ensure_binaries_loaded();
-}
 
 int DeviceRunner::ensure_binaries_loaded() {
     // AICPU .so: load-once, matching onboard's binaries_loaded_ pattern.
     // Keeping the DSO alive across runs preserves g_aicpu_executor state
     // (orch_so_handle_ etc.), which is required for the orch-SO cache-hit path.
     if (!aicpu_so_loaded_ && !aicpu_so_binary_.empty()) {
-        if (!create_temp_so_file(
+        if (!simpler::common::sim_host::create_temp_so_file(
                 "/tmp/aicpu_sim_XXXXXX", aicpu_so_binary_.data(), aicpu_so_binary_.size(), &aicpu_so_path_
             )) {
             LOG_ERROR("Failed to create temp file for AICPU SO");
@@ -242,98 +74,38 @@ int DeviceRunner::ensure_binaries_loaded() {
             return -1;
         }
 
-        aicpu_execute_func_ = reinterpret_cast<int (*)(Runtime *)>(dlsym(aicpu_so_handle_, "aicpu_execute"));
-        if (aicpu_execute_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for aicpu_execute: %s", dlerror());
-            return -1;
-        }
+        auto load_sym = [this](const char *name, void **out) -> bool {
+            void *sym = dlsym(aicpu_so_handle_, name);
+            if (sym == nullptr) {
+                LOG_ERROR("dlsym failed for %s: %s", name, dlerror());
+                return false;
+            }
+            *out = sym;
+            return true;
+        };
 
-        set_platform_regs_func_ = reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_regs"));
-        if (set_platform_regs_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_regs: %s", dlerror());
+        if (!load_sym("aicpu_execute", reinterpret_cast<void **>(&aicpu_execute_func_))) return -1;
+        if (!load_sym("set_platform_regs", reinterpret_cast<void **>(&set_platform_regs_func_))) return -1;
+        if (!load_sym("set_platform_dump_base", reinterpret_cast<void **>(&set_platform_dump_base_func_))) return -1;
+        if (!load_sym("set_dump_tensor_enabled", reinterpret_cast<void **>(&set_dump_tensor_enabled_func_))) return -1;
+        if (!load_sym("set_platform_l2_swimlane_base", reinterpret_cast<void **>(&set_platform_l2_swimlane_base_func_)))
             return -1;
-        }
-        set_platform_dump_base_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_dump_base"));
-        if (set_platform_dump_base_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_dump_base: %s", dlerror());
+        if (!load_sym(
+                "set_platform_l2_swimlane_aicore_rotation_table",
+                reinterpret_cast<void **>(&set_platform_l2_swimlane_aicore_rotation_table_func_)
+            ))
             return -1;
-        }
-        set_dump_tensor_enabled_func_ =
-            reinterpret_cast<void (*)(bool)>(dlsym(aicpu_so_handle_, "set_dump_tensor_enabled"));
-        if (set_dump_tensor_enabled_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_dump_tensor_enabled: %s", dlerror());
+        if (!load_sym("set_l2_swimlane_enabled", reinterpret_cast<void **>(&set_l2_swimlane_enabled_func_))) return -1;
+        if (!load_sym("set_platform_pmu_base", reinterpret_cast<void **>(&set_platform_pmu_base_func_))) return -1;
+        if (!load_sym("set_platform_pmu_reg_addrs", reinterpret_cast<void **>(&set_platform_pmu_reg_addrs_func_)))
             return -1;
-        }
-
-        set_platform_l2_swimlane_base_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_l2_swimlane_base"));
-        if (set_platform_l2_swimlane_base_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_l2_swimlane_base: %s", dlerror());
+        if (!load_sym("set_pmu_enabled", reinterpret_cast<void **>(&set_pmu_enabled_func_))) return -1;
+        if (!load_sym("set_platform_dep_gen_base", reinterpret_cast<void **>(&set_platform_dep_gen_base_func_)))
             return -1;
-        }
-
-        set_platform_aicore_rotation_table_func_ = reinterpret_cast<void (*)(uint64_t)>(
-            dlsym(aicpu_so_handle_, "set_platform_l2_swimlane_aicore_rotation_table")
-        );
-        if (set_platform_aicore_rotation_table_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_l2_swimlane_aicore_rotation_table: %s", dlerror());
+        if (!load_sym("set_dep_gen_enabled", reinterpret_cast<void **>(&set_dep_gen_enabled_func_))) return -1;
+        if (!load_sym("set_scope_stats_enabled", reinterpret_cast<void **>(&set_scope_stats_enabled_func_))) return -1;
+        if (!load_sym("set_platform_scope_stats_base", reinterpret_cast<void **>(&set_platform_scope_stats_base_func_)))
             return -1;
-        }
-
-        set_l2_swimlane_enabled_func_ =
-            reinterpret_cast<void (*)(bool)>(dlsym(aicpu_so_handle_, "set_l2_swimlane_enabled"));
-        if (set_l2_swimlane_enabled_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_l2_swimlane_enabled: %s", dlerror());
-            return -1;
-        }
-
-        set_platform_pmu_base_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_pmu_base"));
-        if (set_platform_pmu_base_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_pmu_base: %s", dlerror());
-            return -1;
-        }
-
-        set_platform_pmu_reg_addrs_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_pmu_reg_addrs"));
-        if (set_platform_pmu_reg_addrs_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_pmu_reg_addrs: %s", dlerror());
-            return -1;
-        }
-
-        set_pmu_enabled_func_ = reinterpret_cast<void (*)(bool)>(dlsym(aicpu_so_handle_, "set_pmu_enabled"));
-        if (set_pmu_enabled_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_pmu_enabled: %s", dlerror());
-            return -1;
-        }
-
-        set_platform_dep_gen_base_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_dep_gen_base"));
-        if (set_platform_dep_gen_base_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_dep_gen_base: %s", dlerror());
-            return -1;
-        }
-
-        set_dep_gen_enabled_func_ = reinterpret_cast<void (*)(bool)>(dlsym(aicpu_so_handle_, "set_dep_gen_enabled"));
-        if (set_dep_gen_enabled_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_dep_gen_enabled: %s", dlerror());
-            return -1;
-        }
-
-        set_scope_stats_enabled_func_ =
-            reinterpret_cast<void (*)(bool)>(dlsym(aicpu_so_handle_, "set_scope_stats_enabled"));
-        if (set_scope_stats_enabled_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_scope_stats_enabled: %s", dlerror());
-            return -1;
-        }
-
-        set_platform_scope_stats_base_func_ =
-            reinterpret_cast<void (*)(uint64_t)>(dlsym(aicpu_so_handle_, "set_platform_scope_stats_base"));
-        if (set_platform_scope_stats_base_func_ == nullptr) {
-            LOG_ERROR("dlsym failed for set_platform_scope_stats_base: %s", dlerror());
-            return -1;
-        }
 
         // Log config travels via the RTLD_GLOBAL HostLogger singleton in
         // libsimpler_log.so — already seeded by simpler_log_init() before the
@@ -355,9 +127,8 @@ int DeviceRunner::ensure_binaries_loaded() {
         aicore_so_path_.clear();
     }
 
-    // Write AICore binary to temp file and dlopen
     if (!aicore_kernel_binary_.empty()) {
-        if (!create_temp_so_file(
+        if (!simpler::common::sim_host::create_temp_so_file(
                 "/tmp/aicore_sim_XXXXXX", aicore_kernel_binary_.data(), aicore_kernel_binary_.size(), &aicore_so_path_
             )) {
             LOG_ERROR("Failed to create temp file for AICore SO");
@@ -371,7 +142,9 @@ int DeviceRunner::ensure_binaries_loaded() {
         }
 
         aicore_execute_func_ =
-            reinterpret_cast<aicore_execute_func_t>(dlsym(aicore_so_handle_, "aicore_execute_wrapper"));
+            reinterpret_cast<void (*)(Runtime *, int, CoreType, uint32_t, uint64_t, uint32_t, uint64_t)>(
+                dlsym(aicore_so_handle_, "aicore_execute_wrapper")
+            );
         if (aicore_execute_func_ == nullptr) {
             LOG_ERROR("dlsym failed for aicore_execute_wrapper: %s", dlerror());
             return -1;
@@ -393,42 +166,16 @@ int DeviceRunner::ensure_binaries_loaded() {
     return 0;
 }
 
-void *DeviceRunner::allocate_tensor(size_t bytes) { return mem_alloc_.alloc(bytes); }
-
-void DeviceRunner::free_tensor(void *dev_ptr) {
-    if (dev_ptr != nullptr) {
-        mem_alloc_.free(dev_ptr);
-    }
-}
-
-int DeviceRunner::copy_to_device(void *dev_ptr, const void *host_ptr, size_t bytes) {
-    // In simulation, this is just a memcpy
-    std::memcpy(dev_ptr, host_ptr, bytes);
-    return 0;
-}
-
-int DeviceRunner::copy_from_device(void *host_ptr, const void *dev_ptr, size_t bytes) {
-    // In simulation, this is just a memcpy
-    std::memcpy(host_ptr, dev_ptr, bytes);
-    return 0;
-}
-
 int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     clear_cpu_sim_shared_storage();
-    // Validate launch_aicpu_num
     if (launch_aicpu_num < 1 || launch_aicpu_num > PLATFORM_MAX_AICPU_THREADS) {
         LOG_ERROR("launch_aicpu_num (%d) must be in range [1, %d]", launch_aicpu_num, PLATFORM_MAX_AICPU_THREADS);
         return -1;
     }
 
     // Validate block_dim. Sim has no stream resource query, so the static
-    // platform capacity is the bound (mirrors the onboard fallback in
-    // DeviceRunner::validate_block_dim). The scheduler assigns cores to
-    // threads cluster-aligned round-robin, so block_dim need not be evenly
-    // divisible by the scheduler thread count.
-    //
-    // block_dim == 0 is the CallConfig "auto" sentinel — resolve to the
-    // static max since sim has no per-stream resource query.
+    // platform capacity is the bound. block_dim == 0 is the CallConfig "auto"
+    // sentinel.
     if (block_dim == 0) {
         block_dim = PLATFORM_MAX_BLOCKDIM;
         LOG_INFO_V0("block_dim auto-resolved to %d", block_dim);
@@ -438,17 +185,15 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         return -1;
     }
 
-    // Ensure device is initialized
     int rc = ensure_device_initialized();
     if (rc != 0) {
         LOG_ERROR("ensure_device_initialized failed: %d", rc);
         return rc;
     }
 
-    // Lazy-allocate the 8-byte device_wall buffer on first run. Sim's
-    // "device pointer" is a host malloc returned by allocate_tensor; the
-    // sim AICPU thread (host CPU thread) writes through this pointer just
-    // like onboard's AICPU does.
+    // Lazy-allocate the 8-byte device_wall buffer on first run. Sim's "device
+    // pointer" is a host malloc returned by allocate_tensor; the sim AICPU
+    // thread writes through this pointer just like onboard's AICPU does.
     if (device_wall_dev_ptr_ == nullptr) {
         device_wall_dev_ptr_ = allocate_tensor(sizeof(uint64_t));
         if (device_wall_dev_ptr_ != nullptr) {
@@ -457,7 +202,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         }
     }
 
-    // Calculate execution parameters
     block_dim_ = block_dim;
     int num_aicore = block_dim * cores_per_blockdim_;
 
@@ -466,12 +210,10 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         return -1;
     }
 
-    // Initialize handshake buffers
     runtime.worker_count = num_aicore;
     worker_count_ = num_aicore;
     runtime.aicpu_thread_num = launch_aicpu_num;
 
-    // Calculate number of AIC cores
     int num_aic = block_dim;
     uint32_t enable_profiling_flag = PROFILING_FLAG_NONE;
     if (enable_dump_tensor_) {
@@ -495,13 +237,9 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         runtime.workers[i].aicpu_ready = 0;
         runtime.workers[i].aicore_done = 0;
         runtime.workers[i].task = 0;
-        // First 1/3 are AIC, remaining 2/3 are AIV
         runtime.workers[i].core_type = (i < num_aic) ? CoreType::AIC : CoreType::AIV;
     }
 
-    // Set function_bin_addr for each task: Runtime::func_id_to_addr_[] stores
-    // a CoreCallable host address (chip buffer + offset); dereference
-    // resolved_addr_ for the dlsym function pointer.
     LOG_DEBUG("Setting function_bin_addr for Tasks (Simulation)");
     for (int i = 0; i < runtime.get_task_count(); i++) {
         Task *task = runtime.get_task(i);
@@ -519,10 +257,8 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         return rc;
     }
 
-    // Store runtime pointer for print_handshake_results
     last_runtime_ = &runtime;
 
-    // Initialize per-subsystem shared memory.
     if (enable_l2_swimlane_) {
         rc = init_l2_swimlane(num_aicore, device_id_);
         if (rc != 0) {
@@ -532,7 +268,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     }
 
     if (enable_dump_tensor_) {
-        // Initialize tensor dump (independent from profiling)
         rc = init_tensor_dump(runtime, device_id_);
         if (rc != 0) {
             LOG_ERROR("init_tensor_dump failed: %d", rc);
@@ -565,14 +300,11 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     }
 
     // On any exit from run() — success or early error — release the diagnostics
-    // collectors' shared memory. They are only re-initialized per run(), so a
-    // Worker reused across runs (e.g. a pytest session-scoped worker pool) would
-    // otherwise re-enter init_l2_swimlane() with stale state still allocated.
+    // collectors' shared memory.
     auto perf_cleanup = RAIIScopeGuard([this]() {
         finalize_collectors();
     });
 
-    // Allocate simulated register blocks for all AICore cores
     size_t total_reg_size = num_aicore * SIM_REG_BLOCK_SIZE;
     void *reg_blocks = mem_alloc_.alloc(total_reg_size);
     if (reg_blocks == nullptr) {
@@ -585,7 +317,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         mem_alloc_.free(reg_blocks);
     });
 
-    // Build array of per-core register base addresses
     size_t regs_array_size = num_aicore * sizeof(uint64_t);
     uint64_t *regs_array = reinterpret_cast<uint64_t *>(mem_alloc_.alloc(regs_array_size));
     if (regs_array == nullptr) {
@@ -609,8 +340,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     // Allocate simulated PMU register blocks. PMU MMIO is a separate address
     // region from the general AICore regs on hardware, so sim mirrors that with
     // its own backing memory; otherwise the AICPU PMU collector would early-out
-    // on every core and the buffer-pool / collector-thread / CSV path would
-    // never run in sim.
+    // on every core.
     size_t total_pmu_reg_size = num_aicore * SIM_REG_BLOCK_SIZE;
     void *pmu_reg_blocks = mem_alloc_.alloc(total_pmu_reg_size);
     if (pmu_reg_blocks == nullptr) {
@@ -623,8 +353,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         mem_alloc_.free(pmu_reg_blocks);
     });
 
-    // physical_core_id == logical i in sim (see Launch AICore threads below),
-    // so a num_aicore-sized array indexed by physical core id is sufficient.
     size_t pmu_regs_array_size = num_aicore * sizeof(uint64_t);
     uint64_t *pmu_regs_array = reinterpret_cast<uint64_t *>(mem_alloc_.alloc(pmu_regs_array_size));
     if (pmu_regs_array == nullptr) {
@@ -645,7 +373,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
 
     LOG_INFO_V0("Allocated simulated PMU registers: %d cores x 0x%x bytes", num_aicore, SIM_REG_BLOCK_SIZE);
 
-    // Check if executors are loaded
     if (aicpu_execute_func_ == nullptr || aicore_execute_func_ == nullptr || set_platform_regs_func_ == nullptr ||
         set_platform_dump_base_func_ == nullptr || set_dump_tensor_enabled_func_ == nullptr ||
         set_platform_pmu_base_func_ == nullptr || set_platform_pmu_reg_addrs_func_ == nullptr ||
@@ -660,7 +387,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     set_platform_dump_base_func_(kernel_args_.dump_data_base);
     set_dump_tensor_enabled_func_(enable_dump_tensor_);
     set_platform_l2_swimlane_base_func_(kernel_args_.l2_swimlane_data_base);
-    set_platform_aicore_rotation_table_func_(kernel_args_.l2_swimlane_aicore_rotation_table);
+    set_platform_l2_swimlane_aicore_rotation_table_func_(kernel_args_.l2_swimlane_aicore_rotation_table);
     set_l2_swimlane_enabled_func_(enable_l2_swimlane_);
     set_platform_pmu_base_func_(kernel_args_.pmu_data_base);
     set_platform_pmu_reg_addrs_func_(kernel_args_.pmu_reg_addrs);
@@ -670,13 +397,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     set_scope_stats_enabled_func_(enable_scope_stats_);
     set_platform_scope_stats_base_func_(kernel_args_.scope_stats_data_base);
 
-    // No per-SO log-config push: HostLogger lives in libsimpler_log.so
-    // (RTLD_GLOBAL singleton) and the AICPU sim SO reads it directly via the
-    // same global lookup.
-
     // Start collector mgmt + poll threads now, just before kernels launch.
-    // Starting earlier wastes CPU on empty queues and risks tripping
-    // ProfilerBase's poll-loop idle-timeout if device-side setup is slow.
     auto thread_factory = [this](std::function<void()> fn) {
         return create_thread(std::move(fn));
     };
@@ -696,7 +417,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         scope_stats_collector_.start(thread_factory);
     }
 
-    // Launch AICPU threads (over-launch for affinity gate)
     constexpr int over_launch = PLATFORM_MAX_AICPU_THREADS_JUST_FOR_LAUNCH;
     LOG_INFO_V0("Launching %d AICPU threads (logical=%d)", over_launch, launch_aicpu_num);
     std::vector<std::thread> aicpu_threads;
@@ -704,11 +424,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     std::atomic<int> aicpu_rc{0};
 
     // Sim "device wall" capture — mirrors the onboard kernel.cpp path via the
-    // device_wall_data_base buffer (sim's "device pointer" is a host malloc'd
-    // uint64). sim_t0 is the host steady_clock equivalent of get_sys_cnt_aicpu
-    // in Init; each thread on exit overwrites the buffer with (now - sim_t0).
-    // Last-writer-wins — sim threads are host CPU threads, variance is well
-    // below benchmark noise.
+    // device_wall_data_base buffer.
     if (kernel_args_.device_wall_data_base != 0) {
         *reinterpret_cast<uint64_t *>(kernel_args_.device_wall_data_base) = 0;
     }
@@ -732,7 +448,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         }));
     }
 
-    // Launch AICore threads
     LOG_INFO_V0("Launching %d AICore thread(s)", num_aicore);
     std::vector<std::thread> aicore_threads;
     for (int i = 0; i < num_aicore; i++) {
@@ -746,7 +461,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         }));
     }
 
-    // Wait for all AICPU and AICore threads to complete
     LOG_INFO_V0("Waiting for threads to complete");
     for (auto &t : aicpu_threads) {
         t.join();
@@ -756,9 +470,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     }
     LOG_INFO_V0("All threads completed");
 
-    // Snapshot the device_wall buffer into device_wall_ns_. Sim's "device
-    // memory" is host memory, so this is a single dereference — no DMA.
-    // Mirrors the post-sync rtMemcpy D2H that the onboard DeviceRunner does.
+    // Snapshot the device_wall buffer into device_wall_ns_.
     device_wall_ns_ = 0;
     if (device_wall_dev_ptr_ != nullptr) {
         device_wall_ns_ = *static_cast<uint64_t *>(device_wall_dev_ptr_);
@@ -772,8 +484,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
 
     // Tear down collectors. stop() joins mgmt then collector in the only safe
     // order (mgmt's final-drain pass into L2 has poll as its consumer).
-    // Diagnostic exports use the per-task `output_prefix_` directory the user
-    // set on CallConfig (CallConfig::validate() enforces non-empty upstream).
     if (enable_l2_swimlane_) {
         l2_swimlane_collector_.stop();
         l2_swimlane_collector_.read_phase_header_metadata();
@@ -810,13 +520,11 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
         scope_stats_collector_.write_jsonl(output_prefix_);
     }
 
-    // Print handshake results at end of run
     print_handshake_results();
 
-    // Close AICore kernel .so now while the process is healthy.
-    // AICPU .so is kept alive (load-once) so that g_aicpu_executor state
-    // (orch_so_handle_ etc.) survives across runs for the orch-SO cache-hit path.
-    // It will be closed in finalize() / unload_executor_binaries().
+    // Close AICore kernel .so now while the process is healthy. AICPU .so is
+    // kept alive (load-once) so that g_aicpu_executor state survives across
+    // runs for the orch-SO cache-hit path.
     if (aicore_so_handle_ != nullptr) {
         dlclose(aicore_so_handle_);
         aicore_so_handle_ = nullptr;
@@ -830,20 +538,6 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     return 0;
 }
 
-void DeviceRunner::print_handshake_results() {
-    if (worker_count_ == 0 || last_runtime_ == nullptr) {
-        return;
-    }
-
-    LOG_DEBUG("Handshake results for %d cores:", worker_count_);
-    for (int i = 0; i < worker_count_; i++) {
-        LOG_DEBUG(
-            "  Core %d: aicore_done=%d aicpu_ready=%d task=%d", i, last_runtime_->workers[i].aicore_done,
-            last_runtime_->workers[i].aicpu_ready, last_runtime_->workers[i].task
-        );
-    }
-}
-
 void DeviceRunner::unload_executor_binaries() {
     if (aicpu_so_handle_ != nullptr) {
         dlclose(aicpu_so_handle_);
@@ -853,7 +547,7 @@ void DeviceRunner::unload_executor_binaries() {
         set_platform_dump_base_func_ = nullptr;
         set_dump_tensor_enabled_func_ = nullptr;
         set_platform_l2_swimlane_base_func_ = nullptr;
-        set_platform_aicore_rotation_table_func_ = nullptr;
+        set_platform_l2_swimlane_aicore_rotation_table_func_ = nullptr;
         set_l2_swimlane_enabled_func_ = nullptr;
         set_platform_pmu_base_func_ = nullptr;
         set_platform_pmu_reg_addrs_func_ = nullptr;
@@ -880,182 +574,7 @@ void DeviceRunner::unload_executor_binaries() {
     }
 }
 
-int DeviceRunner::prepare_orch_so(Runtime &runtime) {
-    // Prepared-callable flow only — bytes were staged at
-    // register_callable time; here we only stamp metadata onto
-    // the runtime and resolve `register_new_callable_id_` from first sighting.
-    const int32_t cid = runtime.get_active_callable_id();
-    if (cid < 0) {
-        LOG_ERROR("prepare_orch_so: no active callable_id; prepared-callable flow required");
-        return -1;
-    }
-    auto it = callables_.find(cid);
-    if (it == callables_.end()) {
-        LOG_ERROR("prepare_orch_so: callable_id=%d not registered", cid);
-        return -1;
-    }
-    const auto &state = it->second;
-    // hbg: orch SO never crosses host/device — clear device-orch metadata
-    // and skip AICPU bookkeeping. See onboard/device_runner.cpp.
-    if (state.host_dlopen_handle != nullptr) {
-        runtime.set_dev_orch_so(0, 0);
-        runtime.set_active_callable_id(cid, /*is_new=*/false);
-        return 0;
-    }
-    const bool first_sighting = aicpu_seen_callable_ids_.insert(cid).second;
-    if (first_sighting) {
-        ++aicpu_dlopen_total_;
-    }
-    runtime.set_dev_orch_so(state.dev_orch_so_addr, state.dev_orch_so_size);
-    runtime.set_active_callable_id(cid, first_sighting);
-    LOG_INFO_V0(
-        "Orch SO prepared cid=%d hash=0x%lx %zu bytes (is_new=%d)", cid, state.hash, state.dev_orch_so_size,
-        first_sighting ? 1 : 0
-    );
-    return 0;
-}
-
-int DeviceRunner::register_callable(
-    int32_t callable_id, const void *orch_so_data, size_t orch_so_size, const char *func_name, const char *config_name,
-    std::vector<std::pair<int, uint64_t>> kernel_addrs, std::vector<ArgDirection> signature
-) {
-    // The AICPU executor reserves `orch_so_table_[MAX_REGISTERED_CALLABLE_IDS]`
-    // (declared in src/common/task_interface/callable_protocol.h) and indexes it by
-    // callable_id; rejecting an out-of-range id here keeps the host and
-    // AICPU sides in sync and avoids an OOB access at run time.
-    if (callable_id < 0 || callable_id >= MAX_REGISTERED_CALLABLE_IDS) {
-        LOG_ERROR("register_callable: callable_id=%d out of range [0, %d)", callable_id, MAX_REGISTERED_CALLABLE_IDS);
-        return -1;
-    }
-    if (orch_so_data == nullptr || orch_so_size == 0) {
-        LOG_ERROR("register_callable: empty orch SO for callable_id=%d", callable_id);
-        return -1;
-    }
-    if (callables_.count(callable_id) != 0) {
-        LOG_ERROR("register_callable: callable_id=%d already registered", callable_id);
-        return -1;
-    }
-
-    const uint64_t hash = simpler::common::utils::elf_build_id_64(orch_so_data, orch_so_size);
-
-    auto buf_it = orch_so_dedup_.find(hash);
-    uint64_t dev_addr = 0;
-    if (buf_it == orch_so_dedup_.end()) {
-        void *buf = mem_alloc_.alloc(orch_so_size);
-        if (buf == nullptr) {
-            LOG_ERROR("register_callable: alloc %zu bytes failed", orch_so_size);
-            return -1;
-        }
-        // Sim shares an address space with the simulated AICPU thread, so a
-        // plain memcpy is the moral equivalent of rtMemcpy on hardware.
-        std::memcpy(buf, orch_so_data, orch_so_size);
-        OrchSoBuffer entry;
-        entry.dev_addr = buf;
-        entry.capacity = orch_so_size;
-        entry.refcount = 1;
-        orch_so_dedup_.emplace(hash, entry);
-        dev_addr = reinterpret_cast<uint64_t>(buf);
-        LOG_INFO_V0("register_callable: hash=0x%lx new buffer %zu bytes", hash, orch_so_size);
-    } else {
-        buf_it->second.refcount++;
-        dev_addr = reinterpret_cast<uint64_t>(buf_it->second.dev_addr);
-        LOG_INFO_V0("register_callable: hash=0x%lx shared buffer (refcount=%d)", hash, buf_it->second.refcount);
-    }
-
-    CallableState state;
-    state.hash = hash;
-    state.dev_orch_so_addr = dev_addr;
-    state.dev_orch_so_size = orch_so_size;
-    state.func_name = (func_name != nullptr) ? func_name : "";
-    state.config_name = (config_name != nullptr) ? config_name : "";
-    state.kernel_addrs = std::move(kernel_addrs);
-    state.signature = std::move(signature);
-    callables_.emplace(callable_id, std::move(state));
-    return 0;
-}
-
-int DeviceRunner::register_callable_host_orch(
-    int32_t callable_id, void *host_dlopen_handle, void *host_orch_func_ptr,
-    std::vector<std::pair<int, uint64_t>> kernel_addrs, std::vector<ArgDirection> signature
-) {
-    if (callable_id < 0 || callable_id >= MAX_REGISTERED_CALLABLE_IDS) {
-        LOG_ERROR(
-            "register_callable_host_orch: callable_id=%d out of range [0, %d)", callable_id, MAX_REGISTERED_CALLABLE_IDS
-        );
-        return -1;
-    }
-    if (host_dlopen_handle == nullptr || host_orch_func_ptr == nullptr) {
-        LOG_ERROR("register_callable_host_orch: null handle/fn for callable_id=%d", callable_id);
-        return -1;
-    }
-    if (callables_.count(callable_id) != 0) {
-        LOG_ERROR("register_callable_host_orch: callable_id=%d already registered", callable_id);
-        return -1;
-    }
-
-    CallableState state;
-    state.host_dlopen_handle = host_dlopen_handle;
-    state.host_orch_func_ptr = host_orch_func_ptr;
-    state.kernel_addrs = std::move(kernel_addrs);
-    state.signature = std::move(signature);
-    callables_.emplace(callable_id, std::move(state));
-    ++host_dlopen_total_;
-    LOG_INFO_V0("register_callable_host_orch: cid=%d (host dlopen #%zu)", callable_id, host_dlopen_total_);
-    return 0;
-}
-
-int DeviceRunner::unregister_callable(int32_t callable_id) {
-    auto it = callables_.find(callable_id);
-    if (it == callables_.end()) {
-        return 0;
-    }
-    CallableState state = std::move(it->second);
-    callables_.erase(it);
-    aicpu_seen_callable_ids_.erase(callable_id);
-
-    if (state.host_dlopen_handle != nullptr) {
-        // hbg: dlclose the host handle; no orch SO refcount to decrement.
-        dlclose(state.host_dlopen_handle);
-        return 0;
-    }
-
-    auto buf_it = orch_so_dedup_.find(state.hash);
-    if (buf_it != orch_so_dedup_.end()) {
-        if (--buf_it->second.refcount <= 0) {
-            mem_alloc_.free(buf_it->second.dev_addr);
-            orch_so_dedup_.erase(buf_it);
-        }
-    }
-    return 0;
-}
-
-bool DeviceRunner::has_callable(int32_t callable_id) const { return callables_.count(callable_id) != 0; }
-
-BindCallableResult DeviceRunner::bind_callable_to_runtime(Runtime &runtime, int32_t callable_id) {
-    auto it = callables_.find(callable_id);
-    if (it == callables_.end()) {
-        LOG_ERROR("bind_callable_to_runtime: callable_id=%d not registered", callable_id);
-        return {-1, nullptr, nullptr, 0};
-    }
-    const auto &state = it->second;
-    for (const auto &kv : state.kernel_addrs) {
-        if (kv.first < 0 || kv.first >= RUNTIME_MAX_FUNC_ID) {
-            LOG_ERROR("bind_callable_to_runtime: func_id=%d out of range", kv.first);
-            return {-1, nullptr, nullptr, 0};
-        }
-        runtime.replay_function_bin_addr(kv.first, kv.second);
-    }
-    runtime.set_device_orch_func_name(state.func_name.c_str());
-    runtime.set_device_orch_config_name(state.config_name.c_str());
-    runtime.set_active_callable_id(callable_id, /*is_new=*/false);
-    return {
-        0, state.host_orch_func_ptr, state.signature.empty() ? nullptr : state.signature.data(),
-        static_cast<int>(state.signature.size())
-    };
-}
-
 int DeviceRunner::finalize() {
-    // Skip if already finalized
     if (device_id_ == -1 && aicpu_so_handle_ == nullptr && aicore_so_handle_ == nullptr) {
         return 0;
     }
@@ -1064,48 +583,12 @@ int DeviceRunner::finalize() {
     // perf_cleanup guard; this is the backstop for the no-run-since-init case.
     finalize_collectors();
 
-    // Release any chip callable buffers uploaded via upload_chip_callable_buffer.
-    // Pool semantics mirror per-fid binaries: never freed until finalize.
-    for (auto &kv : chip_callable_buffers_) {
-        for (void *h : kv.second.dlopen_handles) {
-            if (h != nullptr) dlclose(h);
-        }
-        delete[] kv.second.host_scratch;
-        LOG_DEBUG(
-            "Freed chip callable buffer (sim): chip_dev=0x%lx, size=%zu, hash=0x%lx", kv.second.chip_dev,
-            kv.second.total_size, kv.first
-        );
-    }
-    chip_callable_buffers_.clear();
+    release_callable_state();
 
-    // Release any prepared-callable orch SO buffers callers forgot to drop.
-    for (auto &kv : orch_so_dedup_) {
-        if (kv.second.dev_addr != nullptr) {
-            mem_alloc_.free(kv.second.dev_addr);
-        }
-    }
-    orch_so_dedup_.clear();
-    // hbg path: dlclose any host orch handles callers forgot to unregister.
-    // finalize() is the last chance; Worker.close() does not auto-unregister
-    // each callable_id, so without this loop the host process leaks one
-    // dlopen handle per (re)created Worker — observable in long-running
-    // pytest sessions.
-    for (auto &kv : callables_) {
-        if (kv.second.host_dlopen_handle != nullptr) {
-            dlclose(kv.second.host_dlopen_handle);
-        }
-    }
-    callables_.clear();
-    aicpu_seen_callable_ids_.clear();
-    aicpu_dlopen_total_ = 0;
-
-    // Close executor .so files (typically already closed by run(), this is a safety net)
     unload_executor_binaries();
 
-    // Release the three per-Worker pooled arenas (GM heap, PTO2 SM, optional
-    // trb prebuilt runtime arena — each its own device_malloc). Must precede
-    // mem_alloc_.finalize() so the arenas free through the still-live
-    // allocator, not after it.
+    // Release the three per-Worker pooled arenas. Must precede mem_alloc_.finalize()
+    // so the arenas free through the still-live allocator, not after it.
     gm_heap_arena_.release();
     gm_sm_arena_.release();
     runtime_arena_pool_.release();
@@ -1114,14 +597,12 @@ int DeviceRunner::finalize() {
     cached_runtime_arena_size_ = 0;
 
     // Free the 8-byte device_wall buffer (allocated lazily in run()) before
-    // mem_alloc_.finalize(): free_tensor() routes back through mem_alloc_,
-    // so doing it after finalize would be a use-after-finalize.
+    // mem_alloc_.finalize().
     if (device_wall_dev_ptr_ != nullptr) {
         free_tensor(device_wall_dev_ptr_);
         device_wall_dev_ptr_ = nullptr;
     }
 
-    // Free all remaining allocations
     mem_alloc_.finalize();
     clear_cpu_sim_shared_storage();
 
@@ -1131,95 +612,6 @@ int DeviceRunner::finalize() {
 
     LOG_INFO_V0("DeviceRunner(sim) finalized");
     return 0;
-}
-
-// =============================================================================
-// Chip Callable Buffer Upload (returns host address of ChipCallable header)
-// =============================================================================
-
-uint64_t DeviceRunner::upload_chip_callable_buffer(const ChipCallable *callable) {
-    if (callable == nullptr || callable->child_count() == 0) {
-        return 0;
-    }
-
-    const ChipCallableLayout layout = compute_chip_callable_layout(callable);
-
-    auto it = chip_callable_buffers_.find(layout.content_hash);
-    if (it != chip_callable_buffers_.end()) {
-        LOG_DEBUG(
-            "Chip callable dedup hit (sim): chip_dev=0x%lx, size=%zu, hash=0x%lx", it->second.chip_dev,
-            it->second.total_size, layout.content_hash
-        );
-        return it->second.chip_dev;
-    }
-
-    // Allocate host scratch (host == device in sim). Plain new[] keeps
-    // ChipCallableBuffer::host_scratch ownership symmetric with finalize().
-    auto *scratch = new uint8_t[layout.total_size];
-    std::memcpy(scratch, callable, layout.total_size);
-
-    // Per-child dlopen + dlsym kernel_entry + register pto-sim hooks, then
-    // patch the child's resolved_addr_ to the function pointer. A scope guard
-    // owns scratch and any dlopen'd handles until the success path dismisses
-    // it; every early return unwinds cleanly.
-    std::vector<void *> dlopen_handles;
-    dlopen_handles.reserve(callable->child_count());
-    auto cleanup = RAIIScopeGuard([&]() {
-        for (void *h : dlopen_handles)
-            dlclose(h);
-        delete[] scratch;
-    });
-
-    for (int32_t i = 0; i < callable->child_count(); ++i) {
-        const uint32_t off = callable->child_offset(i);
-        auto *child_in_scratch = reinterpret_cast<CoreCallable *>(scratch + layout.header_size + off);
-        const void *kernel_binary = child_in_scratch->binary_data();
-        size_t kernel_size = static_cast<size_t>(child_in_scratch->binary_size());
-
-        std::string tmpfile;
-        if (!create_temp_so_file(
-                "/tmp/kernel_" + std::to_string(callable->child_func_id(i)) + "_XXXXXX",
-                reinterpret_cast<const uint8_t *>(kernel_binary), kernel_size, &tmpfile
-            )) {
-            LOG_ERROR("Failed to create temp file for child kernel #%d", i);
-            return 0;
-        }
-
-        void *handle = dlopen(tmpfile.c_str(), RTLD_NOW | RTLD_LOCAL);
-        std::remove(tmpfile.c_str());
-        if (!handle) {
-            LOG_ERROR("dlopen failed for child kernel #%d: %s", i, dlerror());
-            return 0;
-        }
-        dlopen_handles.push_back(handle);
-
-        void *func = dlsym(handle, "kernel_entry");
-        if (!func) {
-            LOG_ERROR("dlsym failed for child kernel #%d 'kernel_entry': %s", i, dlerror());
-            return 0;
-        }
-
-        auto register_hooks = reinterpret_cast<void (*)(void *, void *)>(dlsym(handle, "pto_sim_register_hooks"));
-        if (register_hooks != nullptr) {
-            register_hooks(
-                reinterpret_cast<void *>(pto_sim_get_subblock_id),
-                reinterpret_cast<void *>(pto_sim_get_pipe_shared_state)
-            );
-        }
-
-        child_in_scratch->set_resolved_addr(reinterpret_cast<uint64_t>(func));
-    }
-
-    cleanup.dismiss();
-    const uint64_t chip_dev = reinterpret_cast<uint64_t>(scratch);
-    chip_callable_buffers_.emplace(
-        layout.content_hash, ChipCallableBuffer{chip_dev, scratch, layout.total_size, std::move(dlopen_handles)}
-    );
-    LOG_DEBUG(
-        "Uploaded chip callable (sim): chip_dev=0x%lx, size=%zu, child_count=%d, hash=0x%lx", chip_dev,
-        layout.total_size, callable->child_count(), layout.content_hash
-    );
-    return chip_dev;
 }
 
 // =============================================================================
@@ -1234,7 +626,6 @@ int DeviceRunner::init_l2_swimlane(int num_aicore, int device_id) {
         return mem_alloc_.free(dev_ptr);
     };
 
-    // Simulation: dev pointer is directly host-accessible, no register pass-through.
     int rc = l2_swimlane_collector_.initialize(
         num_aicore, device_id, l2_swimlane_level_, alloc_cb, nullptr, free_cb, output_prefix_
     );
@@ -1295,8 +686,6 @@ int DeviceRunner::init_dep_gen(int num_threads, int /*device_id*/) {
         return mem_alloc_.free(dev_ptr);
     };
 
-    // Sim shares an address space with the AICPU thread, so register_cb is
-    // not needed (mirrors PMU's nullptr in sim).
     int rc = dep_gen_collector_.init(num_threads, alloc_cb, nullptr, free_cb, -1);
     if (rc != 0) {
         return rc;
@@ -1314,8 +703,6 @@ int DeviceRunner::init_scope_stats(int num_threads) {
         return mem_alloc_.free(dev_ptr);
     };
 
-    // Sim shares an address space with the AICPU thread, so register_cb is
-    // not needed (mirrors PMU / dep_gen's nullptr in sim).
     int rc = scope_stats_collector_.init(num_threads, alloc_cb, /*register_cb=*/nullptr, free_cb, /*device_id=*/-1);
     if (rc != 0) {
         return rc;
@@ -1327,8 +714,6 @@ int DeviceRunner::init_scope_stats(int num_threads) {
 }
 
 void DeviceRunner::finalize_collectors() {
-    // Free through MemoryAllocator so finalize() can audit. Sim shares an
-    // address space with the AICPU thread, so no host unregister is needed.
     auto free_cb = [this](void *dev_ptr) -> int {
         return mem_alloc_.free(dev_ptr);
     };
