--- src/a5/platform/sim/host/memory_allocator.cpp
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) PyPTO Contributors.
- * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
- * CANN Open Software License Agreement Version 2.0 (the "License").
- * Please refer to the License for details. You may not use this file except in compliance with the License.
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
- * See LICENSE in the root of the software repository for the full text of the License.
- * -----------------------------------------------------------------------------------------------------------
- */
-/**
- * Memory Allocator Implementation (Simulation)
- *
- * Uses standard malloc/free to simulate device memory operations.
- */
-
-#include "host/memory_allocator.h"
-
-#include <cstdlib>
-#include "common/unified_log.h"
-
-MemoryAllocator::~MemoryAllocator() { finalize(); }
-
-void *MemoryAllocator::alloc(size_t size) {
-    void *ptr = std::malloc(size);
-    if (ptr == nullptr) {
-        LOG_ERROR("malloc failed (size=%zu)", size);
-        return nullptr;
-    }
-
-    std::scoped_lock<std::mutex> lk(mu_);
-    ptr_set_.insert(ptr);
-    return ptr;
-}
-
-int MemoryAllocator::free(void *ptr) {
-    if (ptr == nullptr) {
-        return 0;
-    }
-
-    std::scoped_lock<std::mutex> lk(mu_);
-    auto it = ptr_set_.find(ptr);
-    if (it == ptr_set_.end()) {
-        return 0;
-    }
-
-    std::free(ptr);
-    ptr_set_.erase(it);
-    return 0;
-}
-
-int MemoryAllocator::finalize() {
-    std::scoped_lock<std::mutex> lk(mu_);
-    for (void *ptr : ptr_set_) {
-        std::free(ptr);
-    }
-    ptr_set_.clear();
-    return 0;
-}
