+++ src/common/platform/sim/host/memory_allocator.cpp
